images: copy base python files into test container image
We might have useful utilities in our python code. Embed the part that
can be used without ocp-t-f-t into the container image.

At the moment, this code is not very interesting. It will become more
interesting, when we add more utilities.

Note that you can use the container image also for manual testing. In
that case, you can of course just git-clone the repository to get the
python sources. Embedding the python files in the image however has the
benefit that you can run the code directly like:

  $ podman build --no-cache -t quay.io/wizhao/tft-tools:latest -f images/Containerfile .

  $ podman run --rm -ti --name foo --workdir /opt/ocp-tft quay.io/wizhao/tft-tools:latest /opt/pyvenv3.11/bin/python -c 'import host; host.local.run("ps aux")'