images: don' install crictl/cri-tools and no longer enable openSUSE repos
At this moment, download.opensuses.org is not available and breaks
building the container. Regardless of that, it's ugly to include a
non-CentOS-Stream/non-EPEL repository, while there is no cri-tools in
CentOS Stream 9 and EPEL9 as of now.

Also, the correct version of cri-tools depends on the kubernetes
version. While in our case kubernetes does not call crictl from our test
image, it's still unclear whether our crictl safely can interact with
the images on the worker node.

Instead. The worker node already has the right version of crictl
installed.

We no longer use crictl/cri-tools. No longer install it and no longer
enable the openSUSE repositories.