images: don't install crictl/cri-tools and no longer enable openSUSE repos
A few days ago, download.opensuses.org is not available and broke
building the container. Also, it's ugly to use a non-CentOS-Stream,
non-EPEL repository, which is necessary because cri-ctool is not
in CentOS/EPEL.

Also, it's unclear whether the version we install is suitable to
interact with the containers on the worker node.

As the worker node already has the right version of crictl installed, we
can use that instead. We no longer use crictl/cri-tools from the test
image. No longer install it.