k8sClient: add K8sClient.oc_debug() helper
This is a convenience function around calling `kubectl debug` for running
commands on the node.

Unfortunately, we need to explicitly delete the pod afterwards, which
makes this command slow.  In various cases, it can be better to have a
long-running pod that has /host mounted, which then allows to invoke
host commands via `kubectl exec`.

This method has the benefit to not require additional setup (like
starting the container outside). It can be useful in certain cases, or
just to have in the toolbox for some manual (REPL, python -c) testing.

  $ KUBECONFIG=...
    time python -c 'if 1:
        import k8sClient as k

        c = k.K8sClient();
        r = c.oc_debug("echo hi",
                       node_name="worker-247",
                       namespace="default",
                       test_image="quay.io/wizhao/tft-tools:latest")
        print(r)
    '