print_results: fix the script after changing the output format
We discussed, and there is the choice that the output format of the tool
does not need to be stable. That is, it is not guaranteed that you can
parse outputs of a different version of the tool.

Consequently, the output was changed. However, it missed to update
print_results.sh for those changes. Fix it.

Fixes: 1b574558c51c ('evaluator: record TestMetadata in TestResult')