k8sClient: drop usage of unused kubernetes Python library
The kubernetes library was only used for K8sClient.get_nodes() and
K8sClient.get_nodes_with_label(). Those functions were unused.

ocp-tft runs multiple threads, it's important that the lower level
helpers are thread-safe. It's error prone to have some parts of the API
not thread-safe. This could be easily fixed by adding a lock around the
kubernetes calls. But should we do that for unused code?

Already, K8sClient makes the choice to run the `kubectl` binary via
host.Host.run(). I don't particularly care whether we access
functionality via the command line or via the kubernetes Python library.
But using the command line has advantages, because you can see the
command (and the result) in the log file and can easier reproduce by
typing the command. Having it a command also makes the `kubectl` call
(theoretically) remotable via SSH, chroot, sudo, podman-run or a
combination of those.

Let's drop the kubernetes library approach. We can still get the nodes
via `kubectl` and K8sClient.oc(). If that is too cumbersome or
inconvenient, we would add a helper function that does it:

  $ KUBECONFIG=...
    python -c 'if 1:
        import k8sClient

        client = k8sClient.K8sClient()
        nodes = []

        nodes_output = client.oc_get("nodes")
        if nodes_output is not None:
            try:
                nodes = [item["metadata"]["name"] for item in nodes_output["items"]]
            except Exception:
                pass
        print(nodes)
    '