all: use host.Result.success instead of checking returncode
The benefit is that "success" is a concept that we want to check for.
Currently, "success" equals "returncode==0", but it doesn't have to be
the case.

For example, a command to delete something might return with non-zero
exit code but print that the object already does not exist. In that
case, we might want to treat the overall result as a success, despite
the return code. We will add code for that.

In most casees, we should check for "success" (whatever that means in
detail), and not for the "returncode".