task,trafficFlowTests: cleanup logging and handling failures of commands
host.Host.run() already makes an effort to do useful logging. By
default, the command invocation and the result are logged with DEBUG
level. However, there is the option "log_level_fail" to bump the logging
level in case of failure.

Instead, what we previously did is:

    r = self.run_oc(f"apply -f {out_file_yaml}", may_fail=True)
    if not r.success:
        if "already exists" not in r.err:
            logger.error(r)
            sys.exit(-1)

here we make an extra effort to log the result, but it's actually worse than
the logging we could get from Host.run().

Instead, let the automatic logging of host.Host.run() do it's job. For
that, we make use of the arguments "check_success", "may_fail",
"die_on_error", "log_level_fail". We get useful logging for the least
amount of work, without re-implementing the logging.

Notes:

- in _cleanup_previous_testspace(), drop the explicit handling of
  failure. By default we already run with "may_fail=False", which gives
  us a warning level message. That is probably best in this case,
  we could also use "die_on_error=True".

- in _cleanup_previous_testspace(), get rid of various logging messages
  that we start and complete the deletion of the entities. One message
  informing that the cleanup is happening suffices.