host: add "check_success" argument to host.Host.run()
We do

    r = self.run_oc(f"apply -f {out_file_yaml}", may_fail=True)
    if not r.success and "already exists" not in r.err:
        logger.error(r)
        sys.exit(-1)

This means, we need to do a second round of checking, then duplicating
the logging (badly!).

Instead, we can now do

    self.run_oc(f"apply -f {out_file_yaml}",
                check_success=lambda r: r.success or "already exists" in r.err,
                die_on_error=True,
    )

The benefit of this is that the caller does not need to reimplement
error handling and logging. Note how reimplementing it above was done
badly, there is no context about what failed, just blurting out the
output. We can get a better logging with less effort (and not
duplicating the logging logic).