host: make "{host,logger,common,netdev,k8sClient}.py" compatible with Python3.9
CentOS Stream 9 uses Python 3.9. While we can install a 3.11 venv, it is
cumbersome.

In this case, let's still support Python 3.9 for the base modules
"{host,logger,common,netdev,k8sClient}.py".

The rest of ocp-traffic-flow-tests still requires 3.11. The difference
is that we might want to use the base modules at other places, and
requiring multiple users to all use a 3.11 venv is more cumbersome, than
requiring it for ocp-tft.

Changes:

- must use typing.Union[] instead of `|` (Since 3.10)
- no "kw_only" argument to dataclasses (Since 3.10). We use
  KW_ONLY_DATACLASS. This gives us runtime checks on 3.10+
  and makes it clear in code that those classes are supposed
  to be kw_only=True (if we ever upgrade to 3.10, we can find
  all the places that are supposed to use kw_only). What we
  unfortunately lack are static type checks by mypy on 3.10+.
- no typing.assert_type() (Since 3.11)