ktoolbox: move {logger,common,host,netdev,k8sClient}.py to new ktoolbox module
These files contain general purpose utilities, that potentially are
useful everywhere.

Move them to a new module "ktoolbox". This makes it clearer that they are
somehow separate. It also makes it easier to copy+paste the entire
module (or parts of it) to a another project and reuse them via
copy+paste. It also makes it easier to potentially move this code to a
separate git project.

Note that `python /path/to/ocp-tft/ktoolbox/netdev.py` no longer works to
run the main function. You need `(cd /path/to/ocp-tft && python -m
ktoolbox.netdev)` or something similar.