testConfig: ensure TestClient.client() is thread-safe
TestClient for the most part is an immutable object. As such
it is inherently thread-safe.

Once exception is TestClient.client(). This will initialize a K8sClient
on first call. To an external user, this lazy initialization almost
looks as if the client is immutable, because once initialize, it will
not change.

However, the code was not thread-safe. Add a lock for this.