netdev: fix handling non-ethernet addresses in netdev.ip_links()
The address is not always an ethernet address. For example, with
gre type interfaces it is an IPv4 address.

Previously, trying to parse such an address would fail, and
netdev.ip_links() would not return a gre interface.

Instead, reject such addresses and pretend that the address is unset.
This has the benefit, that a user always finds an ethernet address at
this place. The downside is, that the attribute "address" now has a
slightly different meaning than in the output of `ip link`. If we ever
need to know the IP address of gre interfaces, we can rethink this
decision.