netdev: add option to reject reserved names in netdev.normalize_ifname()
Certain interface names conflict with sysctl files created by kernel
modules. That means, in some cases (if the module is not loaded) you are
able to create interfaces with such names, in others not.

Hence, it's a bad idea to use such names when you create an interface.
On the other hand, they may be the name of an interface that is
currently in use, so the functions should accept them at some places.

Add an optional "allow_reserved: bool=True" argument, which can be used
to reject certain well-known names that can conflict.