host/tests: add tests for non-text command line
With errors="surrogateescape", we can also handle non-text output of
commands. But not only that, also the command itself or the arguments
can contain binary data. Granted, as is on Unix, those strings must be
NUL terminated, so it's not arbitrary binary.

This is not something you should ever do. But it's supposed to just
work, if you encode binary data as string with errors="surrogateescape".

Add more tests for that.