host: add log_lineoutput argument to host.Host.run() to show progress
In some cases we have long running commands that print their status
on the output. In that case, it's better to log the result line by line
so we can follow along.

We log the line as repr(line) (where line is in bytes). The upside is
that you can later take the log file and reconstruct the complete
(binary) output. The downside is that it's uglier to read.

  $ python -c 'import ktoolbox.host as h; h.local.run("echo foo; echo err >&2; sleep 0.11; echo foo", log_lineoutput=True)'
  2024-08-21 11:48:23.712 DEBUG   [th:140521504693120]: cmd[1;localhost]: call 'echo foo; echo err >&2; sleep 0.11; echo foo'
  2024-08-21 11:48:23.713 DEBUG   [th:140521504693120]: cmd[1;localhost]: stdout[0] b'foo\n'
  2024-08-21 11:48:23.713 DEBUG   [th:140521504693120]: cmd[1;localhost]: stderr[0] b'err\n'
  2024-08-21 11:48:23.824 DEBUG   [th:140521504693120]: cmd[1;localhost]: stdout[1] b'foo\n'
  2024-08-21 11:48:23.825 DEBUG   [th:140521504693120]: cmd[1;localhost]: └──> 'echo foo; echo err >&2; sleep 0.11; echo foo': success