common: add helper methods for parsing simple type in testConfig
Our input configuration is YAML, and we parse it into a TestConfig
class. The configuration contains many simple types (str, int, bool,
enum), which should be handled in a similar fashion.

Add helper methods for handling such types and use them. The benefit is
that how to handle a str, int, bool, enum is now a common receipt
instead of reimplemented at different places.

There are still few complex properties that require specific code for
parsing and validating them.

Note that various common.structparse_pop_*() methods accept a "default"
or "allow_missing" argument. That controls whether the parameter is
mandatory.  Consequently, also mypy understands whether the function
returns an Optional[T] or a plain T.

Also introduce the common.structparse_with_strdict() context manager.
It's always the same pattern. We first check that the argument is a
dictionary (and shallow-copy it). Then we pop all elements during
parsing, and check that no unknown elements are left at the end. The
first check and the last check can be done via a context manager. The
bad is, that this increases the indentation level by one.

Also, all the structparse_pop_*() functions expect the same first 3
arguments, that would mean a lot of repetition. Instead, we now have a
StructParseVarg type that bundles the [vdict,yamlpath]. Then we can call
varg.for_key(key), which gives us the required 3-tuple to pass to those
functions. It looks a bit odd, but it becomes a recognizable pattern.
Also, mypy correctly understands what's going on, so we don't loose
linter checks and it's hard to misuse.