common: add unwrap() helper
This behaves similar to Rust's unwrap() (hence the name).

It's mainly useful for asserting and unpacking a required/expected
value.

Without it, the linter might complain that we must ensure that the
Optional value is not None. If we expect that to be the case, we can
assert with that with unwrap(), and silence mypy.

This is thus both a runtime assertion and to make the linter happy, in
one.

Note that the generated error message is not very good because there is
no context for a better message (though, you get the backtrace, so it's
not horrible). That is why it's only for assertion use.

With the optional "or_else" parameter, this also provides an unwrap_or().