Update traffic flow test tool to use secondary networks
We need to be able to test out pod2pod traffic on IPUs. Currently the
dpu-operator does not deploy a webhook,so we are modifying the SRIOV pod
yaml to be able to use secondary network nad and also have resource
requests and limits added to the SRIOV pod if the secondary_network_nad
is present. In addition network-attachment-definition no longer by
default should contain the namespace.

In order to use this feature, add secondary_network_nad to the yaml.
resource_name is optional, you can set it if you would like. Otherwise
the program does fetch it for us.  run standard pod to pod traffic on an
IPU cluster. Populate config file as normal.  Pods will properly pass
traffic through the secondary network.  Jira issue:
https://issues.redhat.com/browse/IIC-193