host: let host.Result evaluate in a boolean context
Previously, a bool(host.Result) would always evaluate as True. That is
not useful, especially since Host.run() always returns a Result. So

  res = rsh.run(...)
  if res:
      print("hi")

would always execute. Instead, let Result implement __bool__(), so it
can be used more conveniently.

The use is that currently firewall.nft_call() returns a bool.
In almost all cases, we just want whether there is success
or failure, so the boolean is sufficient. However, internally
we already have a Result at hand. Also, firewall.nft_call()
obviously uses Host.run(), since it has a rsh argument. So
hiding the Result is not useful either. With this change,
we can let the command return a Result, which is still conveniently
usable in a boolean context.