trafficFlowTests: ignore failure to cleanup "multi-networkpolicies"
Only some test cases create the multi-networkpolicies, but during
cleanup we always delete it.

We call client.oc() with may_fail=False, so all failures are logged as
ERROR. That is usually good. However, in this case it's common that the
resource does not exist and we end up logging errors:

  2024-09-16 20:19:03.270 INFO    [th:140347998742336]: Cleaning pods, services and multi-networkpolicies with label tft-tests in namespace default
  2024-09-16 20:19:04.467 ERROR   [th:140347998742336]: cmd[8435;localhost]: └──> 'kubectl --kubeconfig /root/kubeconfig.nicmodecluster -n default delete multi-networkpolicies -l tft-tests': failed (exit 1); err='error: the server doesn\'t have a resource type "multi-networkpolicies"\n'

Detect this case and don't log errors about it.