evalutor: include the bitrate threshold in the test result
the evalutor parses the eval-config.yaml and compares the bitrate
threshold against the bitrate of the test.

While all the data is present, it's in another file.

Let the test result also contain the threshold that was used. Thereby,
if we only look at the result we know which threshold was used.