common: add ExtendedLogger() class
The benefit of an ExtendedLogger instance is that it provides
additional convenience methods (currently only error_and_exit()).

All other Logger methods are delegated to the wrapped logger instance.
As the method error_and_exit() is stateless, there is no difference
between using the extended logger or the wrapped logger.

Hence, with

  ext_logger = common.ExtendedLogger("foo")
  real_logger = logging.getLogger("foo")

calling a method like debug() will yield the same result. That is a
promise.

The benefit here is that the extended logger has convenience methods on
top. While the user does not get access to them via `logging.getLogger()`,
they can create (and cache) an ExtendedLogger instance for that. The
linter will nicely understand when you may call logger.error_and_exit().