testConfig: hook up owner/parent reference of config nodes
TestConfig class parses the YAML configuration file. The main content of
the configuration is in the ConfConfig instance and its nested Conf*
objects. Note that all Conf* objects are immutable, so after the
configuration is parsed, this structure no longer changes.

Clearly we can walk from higher level objects down to the nested ones.
For example, ConfConfig's "tft" attribute contains the deeper nested
ConfTest objects.

Add a way to also walk from the deeper nested objects up to the
parent/owner instance. For that, we use the _owner_reference from
StructParseBase.

Note that the Conf* objects are all immutable and cannot be modified
after being created, to add this circular reference back to the parent,
we need to slightly violate the immutability. This is however solved
nicely by the _owner_reference, which only has an init() and get()
method. The testConfig implementation calls init() to close the loop and
reference the owner. It is enforced that init() can only be called at
most once, so you cannot accidentally reset the instance once it is set.
Also, get() will fail with an exception if the reference is not yet
initialized. The result is that callers can only see the same owner
reference, from their point of view it never changes. For all practical
purposes there is no state here and it is immutable.

Also, internally _owner_reference uses weakref to avoid reference loops
and hamper destruction of objects (e.g. it works without a garbage
collector).