testConfig: cleanup kubeconfigs handling and prefer TestConfig() argument
In non-test code, we currently never don't set the "kubeconfigs"
parameter in the TestConfig() constructor. So the value always either
comes from the YAML configuration or is detected on the file system. The
parameter is currently only used by the tests, so that tests don't fall
back to looking at the file system.

However, previously, the content of the YAML would take precedence over
the "kubeconfigs" argument. As said, this doesn't really matter for now,
but in general it doesn't seem expected. We want that the argument to
the function overwrites the value from the YAML configuration. Change
the precedence and prefer the "kubeconfigs" argument to the TestConfig()
constructor over what is in the YAML config.