task: ignore errors from failed commands in task operation
For running the command, we already ignored the exit code of the
command. That is, because usually the command will be stopped by killing
the process, and it's often that the exit code is non-zero.

For the stop/cancel command, now also ignore the exit code of the
command. Seems there are cases where `killall iperf3` can fail. Note
that it can be that the other side already terminated, which causes the
iperf3 process to quit (--one-off) and killall to fail. Ignore that
condition too.