tests: move "updated-eval-config.yaml" to "tests/eval-config-2.yaml"
This file might be useful for unit testing to have some expected file.
Keep it. But it's not useful to have in the top level.

Note that "generate_new_eval_config.py" can generate this file. But even
if we run that script, we don't want to have the previous file commited
to git.