generate_new_eval_config.py: remove broken script
"generate_new_eval_config.py" no longer works as the file format
changed. That is partly because there are no unit tests and because
the code does not use strong typing so we get no mypy warnings.

This maybe should be fixed, and maybe one day it will. But for
now, this script is just broken. Delete it.

If you want to bring it back, do so and fix the problems.