evaluator: accept missing or empty "eval-config.yaml"
Evaluator takes "eval-config.yaml" for getting the per-test-case
thresholds. It already accepts that  the config only specifies the
thresholds of some test cases. In such case, test cases without a
threshold will pass evaluation successfully.

Extend this, so we can pass in a completely empty YAML or have no config
at all. An empty config is just an extreme case of having incomplete
config.

Now, the YAML file can be empty (which YAML parser parses to `None`) or
it can contain an empty dictionary `{}`. Also, the filename can be empty
too, which is again the same as providing an empty config.

This is for consistently supporting the extreme case of having only a
partial configuration, the one where all configuration is absent.