tft: rename "main.py" to "tft.py"
We have multple python scripts that are directly executable (i.e. they
have a main() function).

When running the test, it's not obvious why this program is called
`main.py`. Rename it to `tft.py`. You should now run `tft.py`.

Add a symlink `main.py` that points to the new script. This is for
backward compatibility, so that the various callers don't need to update
their code (yet).