tftbase: pretty print JSON output when writing TftResults
Pretty printed JSON is much nicer to read. The huge long line is also
problematic with editors and gnu-less.

Also, our sources contain such output files for the unit tests. Having a
huge file in git with only one line is also a problem there. Also, when
the file changes, we cannot see the difference.

Instead, ensure we pretty print the JSON.

Use ktoolbox.common.json_dump() for this, which exists exactly for doing
this the way we want it for this usecase.

Also, update the files in our source tree to be formatted the same way.