evalConfig: make threshold optional and per rx/tx thresholds
Previously, the TestItem.threshold value was mandatory. That is, while
the eval-config may contain only partial configurations, whenever it had
a TestCaseData entry, then both Normal and Reverse settings and their
threshold were required.

That doesn't make too much sense, because most tests cannot even run in
reverse. Also, since we allow for the configuration to contain partial
configuration, and since we already need to handle missing
configuration, we should also allow missing entries and/or missing
thresholds.

Also, the threshold really depends on the direction of the traffic.
Replace the "threshold" fields with "threshold_rx" and "threshold_tx".
That will be useful, because we will want to generate the thresholds
based on previous runs. But then it is not clear how to calculate a
combined threshold based on separate rx/tx measurements.