generate_eval_config: add new tool for generating eval-config thresholds
We had a similar tool, however that bitrot and no longer worked. It was
dropped in [1]. Bring something similar back, but it's also quite
different.

The new tool takes a list of previous runs. It computes a new threshold
and writes an "eval-config.yaml".

It can also take an existing configuration with the "--config" option.
In that case it will keep the entries from the configuration and only
update the thresholds. With "--tigthen-only", it will only increase the
thresholds, not increase them.

This is an initial approach to answer [2] and a first attempt. It
probably needs further tweaking for example in the way how old runs
are combined to new thresholds. Howver, it's added now, so that we have
the general pieces in place and a basis for future improvements.

[1] https://github.com/wizhaoredhat/ocp-traffic-flow-tests/commit/137b00dcfb4025ab23011d21cb0860c73da922d2
[2] https://issues.redhat.com/browse/NHE-774