eval-config: regenerate "eval-config.yaml"
We now have a "generate_eval_config.py" tool that can update an existing
eval-config.yaml.

It is reasonable to start with the eval-config.yaml we have commited in
git. And it's useful to be able to compare (diff) the result of the
generate command with the orignal file. For that, the orginal file
should not be written by hand, but be formatted the same way, so that
the diff is smaller.

Regenerate the file with:

  ./generate_eval_config.py --config eval-config.yaml --output eval-config.yaml

Also, add a unit test to ensure it stays in sync.