github: update tft-test-image container URL after migration
The repository migrated, so the tag/URL of the container needs to be
adjusted.

Also, in the github workflow don't detect the "owner" via
a separate build step. Other places in the repository also hard
code the actual URL, so if we ever move it again, we anyway have
to update those places. Having the owner part detected makes it
harder to understand which tag will be used.

Signed-off-by: Thomas Haller <thaller@redhat.com>