TestSettings: drop redundant "conf_{server,client}" arguments
Redundancy is bad. This value is already obtainable via
"TestSettings.connection". Note that "TestSettings.connection" and
"TestSettings.cfg_descr" are all guaranteed to be immutable. Hence,
looking up the same value multiple times is also guaranteed to yield the
same value. No need to pass around a redundant value.

Even if tomorrow we support multiple Conf{Server,Client} instances in
our configuration, we certainly don't want to pass the
"conf_{server,client}" objects to TestSettings. Instead, we would pass
the indexes, so that we know which one of the many we talk about.

Signed-off-by: Thomas Haller <thaller@redhat.com>