TestSettings: add "conf_server_used" property and replace "node_server_name"
With TestCaseTypInfo.is_same_node, some test cases should run on the
same node. This does not seem implemented correctly at all places.

As a first step, add a "TestSettings.conf_server_used", which is beside
"TestSettings.conf_{server,client}".

Also, returning the whole ConfBaseClientServer instance is more powerful
than a property that returns an attribute of that object. Those that
need the name, should look it up themselves. This is fine, because
ConfBaseClientServer is immutable and stateless. Multiple lookups always
give the same name in any case.

Signed-off-by: Thomas Haller <thaller@redhat.com>